Refactor PostgresDatabase in sadu-queries2 module
Removed and replaced the PostgresDatabase class in sadu-queries2 module's test folder. The refactored file now configures a DataSource instance using the 'DataSourceCreator' and executes SQLUpdates upon creation. In addition, a record 'Database' has been added which allows closing of the container, improving resource management.
